Remove package-lock.json
The package-lock.json file for the react-quiz-component has been deleted. This update means that from this point forward, the dependencies of this project will no longer be strictly pinned by default when installing. Consequently, the project can take advantage of automatically receiving minor and patch updates from its dependencies as they are released.